+++ packages/desktop/src-tauri/src/main.rs
@@ -2,5 +2,5 @@
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
 fn main() {
-    pokertrainer_lib::run()
+    pokerbot_lib::run()
 }